chore: Update SocketProvider to use the correct server IP address
The SocketProvider component in the client/src/components/context/SocketProvider.tsx file was updated to use the correct server IP address (http://192.168.69.95:3000) instead of the localhost address. This change ensures that the client can connect to the server correctly.

Also, the package.json and package-lock.json files in the client directory were updated to include the "uuid" package as a dependency. The "@types/uuid" package was also added as a devDependency in the package.json file.

These changes were made to improve the functionality and compatibility of the application.