chore: Update SocketProvider to use production server URL and add uuid dependency
Update SocketProvider in client/src/components/context/SocketProvider.tsx to use the production server URL (https://lomeet-api.vercel.app/) instead of the localhost address. This change ensures that the client can connect to the server correctly.

Also, update the package.json and package-lock.json files in the client directory to include the "uuid" package as a dependency. The "@types/uuid" package is also added as a devDependency in the package.json file.

These changes improve the functionality and compatibility of the application.